Split upload function and event handlers into separate hook and remove from inline `UploadContext`
