+++ src/components/Upload/UploadContext.tsx
@@ -6,18 +6,15 @@ import React, {
   useEffect,
   useReducer,
 } from 'react';
-import { sleep } from '../../util/sleep';
 import * as actions from './Reducer/uploadActions';
 import { initialState } from './Reducer/uploadInitialState';
 import { uploadReducer } from './Reducer/uploadReducer';
 import * as Types from './Reducer/uploadTypings';
-import {
-  useDeleteFileNodeMutation,
-  useRequestFileUploadMutation,
-} from './Upload.generated';
+import { useRequestFileUploadMutation } from './Upload.generated';
 import { UploadItems } from './UploadItems';
 import { UploadManager } from './UploadManager';
 import { useUploadManager } from './UploadManagerContext';
+import { useUploadFile } from './useUploadFile';
 
 const initialUploadContext = {
   addFilesToUploadQueue: (_: Types.FileInput[]) => {
@@ -34,8 +31,8 @@ export const UploadProvider: FC = ({ children }) => {
   const [state, dispatch] = useReducer(uploadReducer, initialState);
   const { setIsManagerOpen } = useUploadManager();
   const { submittedFiles } = state;
-  const [deleteFile] = useDeleteFileNodeMutation();
   const [requestFileUpload] = useRequestFileUploadMutation();
+  const uploadFile = useUploadFile(dispatch);
 
   const addFilesToUploadQueue = useCallback(
     (files: Types.FileInput[]) => {
@@ -56,103 +53,9 @@ export const UploadProvider: FC = ({ children }) => {
     []
   );
 
-  const setUploadError = useCallback(
-    (queueId: Types.UploadFile['queueId'], errorMessage: string) => {
-      dispatch({
-        type: actions.FILE_UPLOAD_ERROR_OCCURRED,
-        queueId,
-        error: new Error(errorMessage),
-      });
-    },
-    []
-  );
-
   const removeUpload = useCallback((queueId: Types.UploadFile['queueId']) => {
     dispatch({ type: actions.REMOVE_UPLOAD, queueId });
   }, []);
-
-  const handleFileUploadSuccess = useCallback(
-    async (file: Types.UploadFile) => {
-      if (file?.callback && file?.uploadId) {
-        const { callback, queueId, fileName, uploadId } = file;
-        try {
-          await callback(uploadId, fileName);
-          dispatch({
-            type: actions.FILE_UPLOAD_COMPLETED,
-            queueId,
-            completedAt: new Date(),
-          });
-          await sleep(10000);
-          removeUpload(queueId);
-        } catch (error) {
-          setUploadError(file.queueId, 'Post-upload action failed');
-          deleteFile({ variables: { id: file.uploadId } });
-        }
-      }
-    },
-    [deleteFile, removeUpload, setUploadError]
-  );
-
-  // This should be rare, it's just there for completeness.
-  // It only runs if upload is complete but status isn't a "success" status.
-  const handleFileUploadCompleteError = useCallback(
-    (statusText, queueId) => {
-      setUploadError(queueId, statusText);
-    },
-    [setUploadError]
-  );
-
-  // This is the error handler that will actually run on errors
-  const handleUploadError = useCallback(
-    (queueId: Types.UploadFile['queueId']) => {
-      setUploadError(queueId, 'Upload failed');
-    },
-    [setUploadError]
-  );
-
-  const handleFileProgress = useCallback(
-    (queueId: Types.UploadFile['queueId'], event: ProgressEvent) => {
-      const { loaded, total } = event;
-      const percentCompleted = Math.floor((loaded / total) * 1000) / 10;
-      dispatch({
-        type: actions.PERCENT_COMPLETED_UPDATED,
-        queueId,
-        percentCompleted,
-      });
-    },
-    []
-  );
-
-  const uploadFile = useCallback(
-    (uploadFile: Types.UploadFile, url: string) => {
-      const { queueId, file } = uploadFile;
-      const xhr = new XMLHttpRequest();
-
-      xhr.onreadystatechange = function () {
-        if (xhr.readyState === XMLHttpRequest.DONE) {
-          const { status } = xhr;
-          const success = status >= 200 && status < 400;
-          if (success) {
-            handleFileUploadSuccess(uploadFile);
-          } else {
-            handleFileUploadCompleteError(xhr.statusText, queueId);
-          }
-        }
-      };
-
-      xhr.upload.onprogress = handleFileProgress.bind(null, queueId);
-      xhr.upload.onerror = () => handleUploadError(queueId);
-      xhr.open('PUT', url);
-      xhr.setRequestHeader('Content-Type', file.type);
-      xhr.send(file);
-    },
-    [
-      handleFileUploadSuccess,
-      handleFileUploadCompleteError,
-      handleFileProgress,
-      handleUploadError,
-    ]
-  );
 
   const handleFileAdded = useCallback(
     async (file: Types.UploadFile) => {
